requirements.txt
attrs==15.2.0
backports-abc==0.5
beautifulsoup4==4.4.1
blinker==1.3
certifi==2016.9.26
chardet==2.3.0
cryptography==1.2.3
Django==1.9
enum34==1.1.2
html5lib==0.999
httplib2==0.9.1
idna==2.0
ipaddress==1.0.16
lxml==3.5.0
oauthlib==1.0.3
piston-mini-client==0.7.5
psycopg2==2.6.1
pyasn1==0.1.9
pyasn1-modules==0.0.7
pycups==1.9.73
pycurl==7.43.0
PyJWT==1.3.0
pyOpenSSL==0.15.1
pyserial==3.0.1
python-dateutil==2.6.0
pyxdg==0.25
redis==2.10.5
service-identity==16.0.0
singledispatch==3.4.0.3
six==1.10.0
sockjs-tornado==1.0.3
SwampDragon==0.4.2.2
tornado==4.4.2
tornado-redis==2.4.18
Twisted==16.0.0
virtualenv==15.1.0
zope.interface==4.1.3
defer==1.0.6
PAM==0.4.2
django-ckeditor==5.2.1
psycopg==2-2.7.1
